+++ server/models/wechat/Room.ts
@@ -58,7 +58,7 @@ Room.init(
       allowNull: false
     },
     unreadCount: {
-      type: new DataTypes.TINYINT(),
+      type: new DataTypes.SMALLINT(),
       defaultValue: 0
     }
   },
